refactor(blackjack): application.kt 가 view 로직을 처리하도록 구현
